Fixed GUI bugs and removed code duplications; Updated installation procedure with container; Updated TFBoost version and README
