+++ gui/Deconvolution.py
@@ -54,8 +54,8 @@ def selectinput():
     filechosen = filedialog.askopenfile(initialdir="", title="Select a File")
     #filelabel = Label(root, text=filechosen,font = ("Arial",))
     #directory = os.path.split(filechosen)[0] + '/' + os.path.split(filechosen)[1]
-    text1 = Text(decW, state='disabled',font = ("Arial",10), width=50, height=1)
-    text1.place (x=195,y=173,width=200)
+    text1 = Text(decW, state='disabled',font = ("Arial",11), width=50, height=1)
+    text1.place (x=185,y=173,width=200)
     text1.configure(state="normal")
     text1.insert('end', os.path.split(filechosen.name)[1])
     text1.configure(state="disabled")
@@ -66,8 +66,8 @@ def selectoutput():
     filechosen1 = filedialog.askopenfile(initialdir="", title="Select a File")
     #filelabel = Label(root, text=filechosen,font = ("Arial",))
     #directory = os.path.split(filechosen)[0] + '/' + os.path.split(filechosen)[1]
-    text1 = Text(decW, state='disabled',font = ("Arial",10), width=50, height=1)
-    text1.place (x=195,y=213,width=200)
+    text1 = Text(decW, state='disabled',font = ("Arial",11), width=50, height=1)
+    text1.place (x=185,y=213,width=200)
     text1.configure(state="normal")
     text1.insert('end', os.path.split(filechosen1.name)[1])
     text1.configure(state="disabled")
@@ -77,7 +77,7 @@ def selectinputspline():
     filechosen2 = filedialog.askopenfile(initialdir="", title="Select a File")
     #filelabel = Label(root, text=filechosen,font = ("Arial",))
     #directory = os.path.split(filechosen)[0] + '/' + os.path.split(filechosen)[1]
-    text1 = Text(decW, state='disabled',font = ("Arial",10), width=50, height=1)
+    text1 = Text(decW, state='disabled',font = ("Arial",11), width=50, height=1)
     text1.place (x=185,y=663,width=200)
     text1.configure(state="normal")
     text1.insert('end', os.path.split(filechosen2.name)[1])
@@ -113,7 +113,7 @@ def openFileOutput():
     global folder_selected3
     folder_selected3 = filedialog.askdirectory()
     directory = os.path.split(folder_selected3)[0] + '/' + os.path.split(folder_selected3)[1]
-    text1 = Text(decW, state='disabled', width=50, height=1, font = ("Arial",10))
+    text1 = Text(decW, state='disabled', width=50, height=1, font = ("Arial",11))
     text1.place (x=200,y=493)
     text1.configure(state="normal")
     text1.insert('end', directory)
@@ -121,7 +121,7 @@ def openFileOutput():
 
 
 def writeCFG( ):
-    text_file = open("../../etc/deconvolution.cfg", 'w+')
+    text_file = open("../etc/deconvolution.cfg", 'w+')
     text_file.write('\nDeconvolution = true;\n') 
 
     text_file.write('\nInputFileConvorTF =' + '"' + os.path.split(filechosen1.name)[0] + '/' + os.path.split(filechosen1.name)[1] + '";\n')
@@ -135,7 +135,7 @@ def writeCFG( ):
 
     text_file.write('\nInputFileCurrent  = ' + '"' + os.path.split(filechosen.name)[0] + '/' + os.path.split(filechosen.name)[1] + '";\n')
     text_file.write('token1    =' + '"' + token_entry.get() + '";\n')
-    text_file.write('token2    =' + '"' + token_entry.get() + '";\n')
+    text_file.write('token2    =' + '"' + token_entry2.get() + '";\n')
     text_file.write('columnT  = 0;\n')
     text_file.write('columnI  = 1;\n')
     text_file.write('Nsamples =' + nsamplesdrop.get() + ';\n')
@@ -156,9 +156,7 @@ def writeCFG( ):
  
     text_file.close()
 
-    os.chdir('../../build')    
-    subprocess.run(["./deconvolution_tbb", "ls -l"])
-    os.chdir('../gui/wsl')  
+    subprocess.run(["./deconvolution_tbb"], cwd="../build")
 
 
 def run_results ():
@@ -167,15 +165,15 @@ def run_results ():
     
 
     if (varlt.get()==0):
-        data1 = pd.read_csv(os.path.split(filechosen.name)[0] + '/' + os.path.split(filechosen.name)[1],sep='\s+',header=None)
+        data1 = pd.read_csv(os.path.split(filechosen.name)[0] + '/' + os.path.split(filechosen.name)[1],sep=r'\s+',header=None)
         data1 = pd.DataFrame(data1)
-        data2 = pd.read_csv(os.path.split(filechosen1.name)[0] + '/' + os.path.split(filechosen1.name)[1],sep='\s+',header=None)
+        data2 = pd.read_csv(os.path.split(filechosen1.name)[0] + '/' + os.path.split(filechosen1.name)[1],sep=r'\s+',header=None)
         data2 = pd.DataFrame(data2)
 
     if (varlt.get()==1):
-        data1 = pd.read_csv(os.path.split(filechosen.name)[0] + '/' + os.path.split(filechosen.name)[1],sep='\s+',header=None, skiprows=1)
+        data1 = pd.read_csv(os.path.split(filechosen.name)[0] + '/' + os.path.split(filechosen.name)[1],sep=r'\s+',header=None, skiprows=1)
         data1 = pd.DataFrame(data1)
-        data2 = pd.read_csv(os.path.split(filechosen1.name)[0] + '/' + os.path.split(filechosen1.name)[1],sep='\s+',header=None, skiprows=1)
+        data2 = pd.read_csv(os.path.split(filechosen1.name)[0] + '/' + os.path.split(filechosen1.name)[1],sep=r'\s+',header=None, skiprows=1)
         data2 = pd.DataFrame(data2)
 
     x1 = data1[0]
@@ -198,7 +196,7 @@ def run_results ():
 
     plt.legend()
  
-    data3 = pd.read_csv(os.path.split(folder_selected3)[0] + '/' + os.path.split(folder_selected3)[1] +'/'+ str(nameentry.get()),sep='\s+',header=None)
+    data3 = pd.read_csv(os.path.split(folder_selected3)[0] + '/' + os.path.split(folder_selected3)[1] +'/'+ str(nameentry.get()),sep=r'\s+',header=None)
     data3 = pd.DataFrame(data3)
 
     x3 = data3[0]
@@ -250,50 +248,50 @@ lab2.place(x=30,y=120)
 lab2b = Label(decW, text='(the time step is set to 1ps by default)',font = ("Modern",9))
 lab2b.place(x=30,y=140)
 
-btnIN = Button(decW,text='Choose Input file:',command = selectinput,font = ("Arial",10), width=15)
+btnIN = Button(decW,text='Choose Input file:',command = selectinput,font = ("Arial",11), width=15)
 btnIN.place(x=30,y=170)
 
-btnOUT = Button(decW,text='Choose Output file:',command = selectoutput,font = ("Arial",10), width=15)
+btnOUT = Button(decW,text='Choose Output file:',command = selectoutput,font = ("Arial",11), width=15)
 btnOUT.place(x=30,y=210)
 
-lab3 = Label(decW,text='Lines to skip in input file:',font = ("Arial",10))
+lab3 = Label(decW,text='Lines to skip in input file:',font = ("Arial",11))
 lab3.place(x=30,y=260)
 
-LTskip = Entry(decW,font = ("Arial",10))
+LTskip = Entry(decW,font = ("Arial",11))
 LTskip .place(x=200,y=260,width=40)
 LTskip .insert(END,'0')
 
-lab3b = Label(decW,text='Lines to skip in output file:',font = ("Arial",10))
-lab3b.place(x=290,y=260)
+lab3b = Label(decW,text='Lines to skip in output file:',font = ("Arial",11))
+lab3b.place(x=300,y=260)
 
-LTskip2 = Entry(decW,font = ("Arial",10))
+LTskip2 = Entry(decW,font = ("Arial",11))
 LTskip2 .place(x=470,y=260,width=40)
 LTskip2 .insert(END,'0')
 
-lab3c = Label(decW,text='LTSpice signals:',font = ("Arial",10))
+lab3c = Label(decW,text='LTSpice signals:',font = ("Arial",11))
 lab3c.place(x=535,y=260)
 varlt = IntVar()
-Checkbutton(decW, variable=varlt,command=ltspiceDEC,font = ("Arial",10)).place(x=645, y=258)
+Checkbutton(decW, variable=varlt,command=ltspiceDEC,font = ("Arial",11)).place(x=645, y=258)
 
-lab4 = Label(decW,text='Scale factor:',font = ("Arial",10))
+lab4 = Label(decW,text='Scale factor:',font = ("Arial",11))
 lab4.place(x=30,y=290)
 
-scaleF = Entry(decW,font = ("Arial",10))
+scaleF = Entry(decW,font = ("Arial",11))
 scaleF .place(x=200,y=290,width=40)
 scaleF .insert(END,'1')
 
 
 #choose the token#################################
-lab5 = Label(decW,text='Token1:',font = ("Arial",10))
+lab5 = Label(decW,text='Token1:',font = ("Arial",11))
 lab5.place(x=30,y=320)
 
 #choose the token of the output file #################################
-lab5b = Label(decW,text='Token2:',font = ("Arial",10))
+lab5b = Label(decW,text='Token2:',font = ("Arial",11))
 lab5b.place(x=350,y=320)
 
-tokendrop = ttk.Combobox(decW, width = 10, font=("Arial",10))
+tokendrop = ttk.Combobox(decW, width = 10, font=("Arial",11))
 
-tokendrop2 = ttk.Combobox(decW, width = 10, font=("Arial",10))
+tokendrop2 = ttk.Combobox(decW, width = 10, font=("Arial",11))
   
   
 # Adding combobox drop down list
@@ -328,12 +326,12 @@ tokendrop2.bind("<<ComboboxSelected>>", tokenget)
 
 
 # set Nsamples #################################
-lab6 = Label(decW,text='Nsamples:',font = ("Arial",10))
+lab6 = Label(decW,text='Nsamples:',font = ("Arial",11))
 lab6.place(x=30,y=350)
 
 n1 = StringVar()
 nsamplesdrop = ttk.Combobox(decW, width = 10,
-                            textvariable = n1, font=("Arial",10))
+                            textvariable = n1, font=("Arial",11))
   
 # Adding combobox drop down list
 nsamplesdrop['values'] = ('8192', 
@@ -348,39 +346,39 @@ nsamplesdrop.current(2)
 
 nsamplesdrop.place(x=200,y=350)
 
-lab6b = Label(decW,text='dT:',font = ("Arial",10))
+lab6b = Label(decW,text='dT:',font = ("Arial",11))
 lab6b.place(x=380,y=350)
 
-dTdec = Entry(decW,font = ("Arial",10))
+dTdec = Entry(decW,font = ("Arial",11))
 dTdec .place(x=415,y=350,width=70)
 dTdec .insert(END,'1e-12')
 
 #POSTPROCESSING FILTER
-my_label6 = Label(decW, text='Filter transfer function:',font = ("Arial",10))
+my_label6 = Label(decW, text='Filter transfer function:',font = ("Arial",11))
 my_label6.place(x=30,y=380)
 var13 = IntVar()
-Checkbutton(decW, variable=var13,font = ("Arial",10)).place(x=192, y=380)
+Checkbutton(decW, variable=var13,font = ("Arial",11)).place(x=192, y=380)
 var13.set(1)
 
-my_label7 = Label(decW, text='Filter Order n°:',font = ("Arial",10))
+my_label7 = Label(decW, text='Filter Order n°:',font = ("Arial",11))
 my_label7.place(x=30,y=410)
-filterN_entry = Entry(decW,font = ("Arial",10))
+filterN_entry = Entry(decW,font = ("Arial",11))
 filterN_entry.place(x=200,y=410,width=40)
 filterN_entry.insert(END,'10')
 
-my_label8 = Label(decW, text='Filter cut-off Frequency:',font = ("Arial",10))
+my_label8 = Label(decW, text='Filter cut-off Frequency:',font = ("Arial",11))
 my_label8.place(x=30,y=440)
-filterF_entry = Entry(decW,font = ("Arial",10))
+filterF_entry = Entry(decW,font = ("Arial",11))
 filterF_entry.place(x=200,y=440,width=40)
 filterF_entry.insert(END,'10e9')
 
-my_button2 = Button(decW, text="Set output folder:",command=openFileOutput,font = ("Arial",10), width=15)
+my_button2 = Button(decW, text="Set output folder:",command=openFileOutput,font = ("Arial",11), width=15)
 my_button2.place(x=30,y=490)
 
-nameLab = Label(decW, text='Set output Filename:',font = ("Arial",10))
+nameLab = Label(decW, text='Set output Filename:',font = ("Arial",11))
 nameLab.place(x=30,y=530)
 
-nameentry = Entry(decW,font = ("Arial",10))
+nameentry = Entry(decW,font = ("Arial",11))
 nameentry.place(x=200,y=530)
 nameentry.insert(END,'filename.txt')
