+++ gui/TFB_guiFiles/csa_eq.png
