+++ gui/TFB_guiFiles/eq_background.png
