+++ gui/TFB_guiFiles/singlestage_eq.png
