--- gui/wsl/TFB_guiFiles/stat2.cfg
@@ -1,336 +1,0 @@
-TIA_BJT_1stage :{
-    ID          = 3;
-    Nsamples    = 4096;
-    Nbins       = 50;
-    minplot     = 0.0;
-    maxplot     = 4e-9;
-    dT          = 1e-12;
-    LEthr       = 0.01;
-    CFD_fr      = 0.35;
-    sigma_noise = 0.00225;
-    r_rednoise  = 0.985112;
-    sampling_dT = 50e-12;
-    bound_fit   = 20;
-    RM_delay    = 100e-12;
-    Rp          = 400.0; 
-    Rf          = 4000.0; 
-    Rl          = 35.0;
-    gm          = 0.615;
-    ci          = 2e-12; 
-    co          = 4e-12; 
-    cf          = 60e-15;
-};
-
-
-TIA_BJT_2stages_GM:{
-    ID          = 2;
-    Nsamples    = 4096;
-    Nbins       = 50;
-    minplot     = 0.0;
-    maxplot     = 4e-9;
-    dT          = 1e-12;
-    LEthr       = 0.01;
-    CFD_fr      = 0.35;
-    sigma_noise = 0.004;
-    r_rednoise  = 0.970446;
-    sampling_dT = 50e-12;
-    bound_fit   = 20;
-    RM_delay    = 100e-12;
-    Rp          = 300.0;
-    Rf          = 3000.0;
-    Rl          = 25.0;
-    gm1         = 0.8;
-    Rf2         = 4000.0;
-    Rl2         = 25.0;
-    gm2         = 0.7;
-    ci          = 1.5e-12;
-    co          = 1.5e-12;
-    co2         = 1.8e-12;
-    cf          = 60e-15;
-};
-
-
-TIA_MOS:{
-    ID          = 0;
-    Nsamples    = 131072;
-    Nbins       = 50;
-    minplot     = 0.0;
-    maxplot     = 50e-9;
-    dT          = 1e-12;
-    LEthr       = 0.03;
-    CFD_fr      = 0.35;
-    sigma_noise = 0.0017;
-    r_rednoise  = 0.998501;
-    sampling_dT = 10e-12;
-    bound_fit   = 20;
-    RM_delay    = 1000e-12;
-    Rf          = 3000000.0;
-    Rl          = 570000.0;
-    gm          = 55e-6;
-    ci          = 105e-15;
-    co          = 21e-15;
-    cf          = 5e-15;
-};
-
-
-TIA_IdealInt:{
-    ID          = 4;
-    Nsamples    = 131072;
-    Nbins       = 50;
-    minplot     = 0.0;
-    maxplot     = 4e-9;
-    dT          = 1e-12;
-    LEthr       = 0.02;
-    CFD_fr      = 0.35;
-    sigma_noise = 0.0017;
-    r_rednoise  = 0.998501;
-    sampling_dT = 50e-12;
-    bound_fit   = 200;
-    RM_delay    = 100e-12;
-    z0          = 110e+3;
-    tau1        = 7.74e-9;
-    tau2        = 18.81e-9;
-    tauz        = 12.66e-9;
-};
-
-FromFile :{
-    ID          = 5;
-    TFFile      = "/home/osboxes/Desktop/TFBoost/build/results_deconv_diamond/TF1ps.txt";
-    //TFFile      = "/home/osboxes/Desktop/TFBoost/build/results_deconv_diamond/kansasTF.txt";
-    NlinesToSkip = 0;
-    Nsamples    = 20000;
-    Nbins       = 100;
-    minplot     = 0.0;
-    maxplot     = 8e-9;
-    dT          = 1e-12;
-    LEthr       = 0.01;
-    CFD_fr      = 0.35;
-    sigma_noise = 0.00281; #0.00225
-    r_rednoise  = 0.998;   #0.985112
-    sampling_dT = 20e-12;
-    bound_fit   = 3;
-    RM_delay    = 200e-12;
-};
-
-
-
-HistogramsPars:{
-
-    TOA_LE:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time of arrival Leading Edge";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    TOA_CFD:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time of arrival CFD";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    TOA_RM:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time of arrival Ref.Method";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    Tpeak:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time at peak";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    Vpeak:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V at peak";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    dvdt_LE:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Slope (dV/dT) on LE threshold";
-        xtitle = "Slope (uV/ps)";
-        ytitle = "counts";
-    };
-    
-    dvdt_CFD:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Slope (dV/dT) on CFD threshold";
-        xtitle = "Slope (uV/ps)";
-        ytitle = "counts";
-    };
-    
-    dvdt_RM:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Slope (dV/dT) on RM threshold";
-        xtitle = "Slope (uV/ps)";
-        ytitle = "counts";
-    };
-    
-    VonThLE:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V on LE threshold";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    VonThCFD:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V on CFD threshold";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    VonThRM:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V on RM threshold";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    TOT:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time over threshold";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-
-    TOA_LE_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time of arrival Leading Edge with noise";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    TOA_CFD_noise:{
-        min    = 0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time of arrival CFD with noise";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    TOA_RM_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time of arrival Ref.Method with noise";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    Tpeak_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time at peak with noise";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-    
-    Vpeak_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V at peak with noise";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    dvdt_LE_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Slope (dV/dT) on LE threshold with noise";
-        xtitle = "Slope (uV/ps)";
-        ytitle = "counts";
-    };
-    
-    dvdt_CFD_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Slope (dV/dT) on CFD threshold with noise";
-        xtitle = "Slope (uV/ps)";
-        ytitle = "counts";
-    };
-    
-    dvdt_RM_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Slope (dV/dT) on RM threshold with noise";
-        xtitle = "Slope (uV/ps)";
-        ytitle = "counts";
-    };
-    
-    VonThLE_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V on LE threshold with noise";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    VonThCFD_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V on CFD threshold with noise";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    VonThRM_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "V on RM threshold with noise";
-        xtitle = "Voltage (V)";
-        ytitle = "counts";
-    };
-    
-    TOT_noise:{
-        min    =  0.0;
-        max    = -1.0;
-        Nbins  = 2000;
-        title  = "Time over threshold with noise";
-        xtitle = "Time (s)";
-        ytitle = "counts";
-    };
-};
-
