--- install/TFBoost_installer
@@ -1,109 +1,0 @@
-#!/bin/bash
-#
-# install TFBoost
-
-
-rc=1
-while [ $rc -eq 1 ]; do
-ans=$(zenity --info --title 'TFBoost Installer GUI' --width=600 --height=400 \
---text 'Clone the repositories of Hydra and TFBoost from git \n (folders will be saved in the current dyrectory)' \
---ok-label Quit \
---extra-button "CLONE" \
---window-icon=logoico.png
-)
-
-rc=$?
-echo "${rc}-${ans}"
-echo $ans
-if [[ $ans = "CLONE" ]]
-
-then
-
-  echo "# Cloning HYDRA"
-  git clone https://github.com/MultithreadCorner/Hydra.git Hydra
-  echo "# Cloning TFBoost"
-  git clone https://github.com/gianmatteocossu/TFBoost.git TFBoost
-  echo "# Setting ROOT CERN path"
-  
-  rp=1
-  while [ $rp -eq 1 ]; do
-	ans1=$(zenity --info --title 'TFBoost Installer GUI' --extra-button "PATH TO ROOT FOLDER" \
-  --text 'ROOT CERN needs to be already installed in this machine.\n Set the path choosing the correct folder'\
-  --width=600 --height=400\
-  --window-icon=logoico.png
-	)
-	
-  rp=$?
-  echo "${rp}-${ans1}"
-  echo $ans1
-  if [[ $ans1 = "PATH TO ROOT FOLDER" ]]
-  then
-	ans3=$(zenity --file-selection --directory --filename '$PWD')
-  sleep 1
-  echo $ans3
-  (
-  
-  echo "Install...";
-
-  echo "25"
-  echo "#Setting enveironment variables..";
-  sleep 1
-
-  export CC=/usr/bin/gcc-8
-  echo "export CC=/usr/bin/gcc-8"
-  export CXX=/usr/bin/g++-8
-  echo "export CXX=/usr/bin/g++-8"
-  export ROOTSYS=$ans3
-  echo "export ROOTSYS=$ans3"
-  export HYDRA_INCLUDE_DIR=$PWD/Hydra
-  echo "export HYDRA_INCLUDE_DIR=$PWD/Hydra"
-
-  echo "30"
-  echo "#INSTALL ALL DEPENDENCIES...insert sudo password";
-  sleep 1
-  sudo apt-get install libconfig-dev libtclap-dev libtbb-dev gcc-8 python3.8 python3-matplotlib python3-tk python3-pip python3-pil.imagetk
-  sudo -H pip3 install pandas
-
-  echo "50"
-  echo "#CONFIGURE TFBoost...";
-  sleep 1
-  cd TFBoost
-  #echo "current directory=$PWD"
-  mkdir -p build
-  cd build
-  #echo "current directory=$PWD"  
-  cmake -DHYDRA_INCLUDE_DIR=$HYDRA_INCLUDE_DIR ../
-  echo "75"
-  echo "#Configuration DONE!";
-  sleep 1
-  echo "80"
-  echo "#Build all Sources: \n\n\n- make analysis_tbb \n\n-deconvolution_tbb \n\n-resampling_tbb \n\n-3Ddiamond_tbb\n\n\n "
-  make analysis_tbb deconvolution_tbb resampling_tbb 3Ddiamond_tbb
-
-  echo "100"
-  echo "# ALL SET!\n\n\nRUN TFBoost with the GUI in folder""TFBoost GUI""\n\ntype: python3 TFBoostGui.py\n\n\n";
-  sleep 10
-
-  )| zenity --progress \
-  --title="TFBoost Installer GUI" \
-  --width=600 \
-  --height=400 \
-  --window-icon=logoico.png
-
-  fi
-  rp=0
-  done
-
-
-fi
-rc=0
-done
-
-echo '' >> ~/.bashrc
-echo 'TFB () {   
-   cd ' ${PWD}'/TFBoost/gui/linux
-   python3 TFBoostGui.py
-}
-' >> ~/.bashrc
-
-
