--- install/installRoot
@@ -1,14 +1,0 @@
-#!/bin/bash
-
-sudo apt-get install gfortran libpcre3-dev \
-xlibmesa-glu-dev libglew1.5-dev libftgl-dev \
-libmysqlclient-dev libfftw3-dev libcfitsio-dev \
-graphviz-dev libavahi-compat-libdnssd-dev \
-libldap2-dev python-dev libxml2-dev libkrb5-dev \
-libgsl0-dev
-
-wget https://root.cern/download/root_v6.24.02.Linux-ubuntu20-x86_64-gcc9.3.tar.gz
-tar -xzvf root_v6.24.02.Linux-ubuntu20-x86_64-gcc9.3.tar.gz
-source root/bin/thisroot.sh 
-echo 'source' $PWD'/root/bin/thisroot.sh' >> ~/.bashrc
-
