--- install/packages~
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-sudo apt-get update
-sudo apt-get upgrade
-sudo apt-get install git libconfig-dev libtclap-dev libtbb-dev gcc-8 python3.8 python3-matplotlib python3-tk python3-pip python3-pil.imagetk cmake g++-8 libgsl-dev libfftw3-dev libconfig++-dev clang-6.0 zenity
-sudo -H pip3 install pandas
